requirements.txt
Flask
numpy
opencv-python
paho-mqtt
